Remove KnightChrMgr & SharedMem. Both are entirely useless to us.

+++ Client/N3Base/AVIPlayer.h
@@ -19,7 +19,7 @@ public:
 
 public:	
 	bool PlayCutScene(LPTSTR pszMovie, HINSTANCE hInstance);
-	static int32_t WINAPI WindowProc_Player( HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam );
+	static LONG WINAPI WindowProc_Player( HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam );
 	static bool		m_bInterrupted;
 protected:
 	HWND			m_hWndMain;
